+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
 
+* 0.1.3
+  - Added Windows Support
+  - Let the user provide the installation directory if the default is not found(windows)
+
+* 0.0.8 
+  - Let the user provide the installation directory if the default is not found(linux)
+
 * 0.0.2 
   - Added Windows(test) support